Preserve the color of ambient light while doing desaturating fogs, fixes #30
The cause of the problem was that we wasn't multiplying the desaturated color by the color of ambient light, which was (unintentionally) equivalent to assuming the light color was always white. This means particles receiving non-white light was emitting more energy than what they received, violating the law of conservation of energy.